backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Authentication
bcrypt==4.1.1
PyJWT==2.8.0
python-jose[cryptography]==3.3.0

# Payment Processing
stripe==7.4.0

# AI Integration
openai==1.3.7
numpy==1.26.2

# AWS Services
boto3==1.29.7

# Caching
redis==5.0.1

# Email
aiosmtplib==3.0.1
email-validator==2.1.0

# Utilities
python-dotenv==1.0.0
httpx==0.25.2
requests==2.31.0
python-json-logger==2.0.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
